+++ src/widgets/planner-edit/index.ts
@@ -1,3 +1,9 @@
 export { resetDragState } from "./lib/resetDragState";
 export { ExcludedTasksSheet } from "./ui/ExcludedTasksSheet";
 export { useExcludedTasksSheetState } from "./model/useExcludedTasksSheetState";
+export {
+  usePlannerEditDnD,
+  type PlannerEditDraggedTask,
+  type PlannerEditInsertPreview,
+  type PlannerEditPreviewSlot,
+} from "./model/usePlannerEditDnD";
